generalize install daemonset
This removes kata-specific details from the Daemonset implementation and
instead makes these configurable via the CRD.

To achieve this a lot of new fields in the CRD are added. It is modeled
after how kata-deploy works, i.e. there is a install and a cleanup
command, a set of environment variables and a label that is added to a
node for when install/uninstall is finished

I let the install daemonset running after the initial installation.
This way, when the cluster is scaled up new nodes that have the correct
labels (NodeSelector is configured via the CRD as well) also get the
runtime installed.

When the CRD is deleted the uninstall daemonset will create pods on all nodes that
had the runtime installed before and run the uninstall scripts of the
uninstall daemonset.

The uninstall daemonset is running in the background but is only
activated when the CR is deleted. The activation works via setting
cleanup labels on the desired nodes. Once these labels are added to
nodes the daemonset will create pods that perform the uninstallation.

Fixes: https://github.com/confidential-containers/operator/issues/26
Signed-off-by: Jens Freimann <jfreimann@redhat.com>