watch for changes to nodes
This adds a Watch for Node changes, which means we get control
every time something is done to a node. When we get control a request
for the reconcile loop is queued.

We need this so that we get control when the Daemonset that uninstalls
the runtime on the node adds/removes a label to the node.

Signed-off-by: Jens Freimann <jfreimann@redhat.com>