controller: Remove kata-cc runtimeClass
We're going to use `kata` as the default runtimeClass for CC,
`kata-qemu` which is the same as `kata`, and we may end up adding a new
one for Cloud Hypervisor.  But, for now, let's remove the `kata-cc` one
as it may actually be confusing to folks as the Kata Containers payload
consumed here is already a specific one for the Confidential Containers
effort.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>