bundle,config: Remove CRI-O mount points for now
Right now CRI-O is not supported and when the time comes that it becomes
supported, reverting this commit is a quite easy task.

For now, let's remove it as an attempt to make it aligned with what
we're doing as part of the Kata Containers scripts for the Confidential
Containers effort.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>