bundle,config: Pin a specific tag for the pre/post images
Instead of relying on the `latest` tag, let's make sure we use a
specific tag for the preInstall and postUninstall images.

This will ensure we don't run into unexpected issues when the image gets
updated, giving us more control on having a functional Operator.

The tag chosen is `2022070719511657216287`, which was tested locally and
contains a build based on the v1.6.6 containerd release, with the needed
patches atop of that.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>