delete labels and pre-install daemonset when CR is being deleted
Currently the labels we use to mark nodes as done are not removed
after the task is completed. To be able to perform the install/uninstall
cycle multiple times we need to reset this status.

This is fixed by removing the labels from the nodes as the last step when the custom
ressource is deleted by the user.

Fixes: #54
Signed-off-by: Jens Freimann <jfreimann@redhat.com>