delete preinstall daemonset when CR is deleted to allow re-installation
We let the preinstall daemonset run even after it is completed so that
the cluster can be scaled by adding/removing nodes. However when the
user deletes the CR we need to remove even the preinstall DS. Otherwise
the preinstall scripts won't be run when a new CR is created.

This patch adds code to delete the preinstall DS as part of the
uninstallation.

Fixes: #50
Signed-off-by: Jens Freimann <jfreimann@redhat.com>