ci: bootstrap the test env and install the operator
Adds ansible playbooks to setup the test environment. Kubernetes will be
deployed with kubeadm.

The cluster/up.sh is used to bring up the cluster. Afterwards run
build_install_operator.sh to build and install the operator and
ccruntime. Notice it pushes the operator's controller image to a docker
registry running locally so we avoid pushing/pulling to/from quay.io.

I tested in a Fedora 34 x86_64 machine in a Vagrant VM (see the
Vagrantfile added in this commit). If you wish to run likewise:

$ cd tests/e2e
$ vagrant up tests-e2e-ubuntu2004

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>