ci: introduce an undo mechanism for bare-metal
When executed on bare-metal machines, most of the changes applied by the
ci scripts should be reverted so that it will leave the system as close
as possible to the initial state and allow subsequent executions.

This introduce a 'undo' mechanism which will attempt to revert the
changes made by the scripts. You should pass '-u' to run-local.sh as
`run-local.sh -u` to enable such as feature.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>