tests/operator.sh: add debuggers
The operator.sh might fail when it is waiting for pods to become
ready and this usually indicates a bug on the operator. However, when
running on CI, it is hard to determine what drove to the pod not coming
fully up. So this added a few debugging instructions for printing
information about the pods for when the operator.sh fail.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>