bundle,config: Update pre-install / post-uninstall image
A new image of container-engine-for-cc-payload (commit 4892b72 from
https://github.com/fidencio/container-engine-for-cc-payload) was created
to fix the post-uninstall pod that is getting stuck on terminating
phase. It turned out that kubelet should not be restarted by the
post-uninstall pod, otherwise it gets in a state where the pod's volumes
cannot be umounted (hence it stays on terminating state forever).

Fixes #53
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>