tests/run-local.sh: undo only the executed steps
The tests/run-local.sh script has the ability to "undo" the instalation
of software and configuration in the local machine. That's usually
needed for CI on bare-metal machine where it needs to revert the
changes made by the job to ensure that the next start on a clean
environment.

This changed the scripts so that it will try to "undo" only the steps
that executed. For example, suppose it has A, B, and C steps and the
execution failed on B; it makes no sense to undo changes on C that never
happened.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>