reconcile: use configurable namespace for the secondary resources
CcRuntime is a cluster scoped resource and therefore their instances
do not specify any namespace info.

However, the secondary resources are namespaced. Currently, a hard-
code value is used but users may want to use some other namespace.

Make the secondary resources namespace configurable using a flag
and env var to the operator. By default, use the same (but have it
dynamic) namespace as what the operator itself runs in.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>