Import container-engine-for-cc-payload
Let's import all the files from the container-engine-for-cc-payload repo
into the Operator one.

The reason for that is for simplifying development and unifying the
components used by the Operator.

I've taken the decision to *not* keep the commit history as that was a
very short one and not worthwhile, IMHO.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>