DO-NOT-MERGE | tests: Enable all the Kata Containers tests
We've been running a subset of the tests running on the Kata Containers
side, which made us not realise that some features (such as measured
boot) were not working with the Operator.

Let's try to enable all the tests, take a look on what breaks, and from
there re-evaluate what needs to be in and what needs to be out.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>