pre-install-payload: allow to pass extra flags to docker manifest
When building the pre-install-payload image for CI it needs to pull/push
the image from a local registry that is not protected. The `docker
manifest` commands (e.g. create) refuses to connect in an unsecure
registry by default, therefore the pre-install-payload build fail. That
can be solved by passing the --insecure flag to `docker manifest` thus
this change allow to pass extra flags to that command.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>