crd: Expose "UsingNFD" to the CcInstallConfig
Let's have a explicit toggle that allows users to specify whether
they're relying on NFD or not.

In case they're not relying on NFD, business as usual.  In case they
are, then the runtime payload will be able to create a specific
NodeFeatureRule, adapt the podOverhead of the runtime class, and make
sure the keys book-keeping is correctly done.

Right now, on the Kata Containers side of the things, only TDX is
capable of doing so, but it's easy to expand in case other TEEs want to
do the same.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>