e2e: Override correct container image
The 'kustomize edit set image' step needs to specify the correct 'name' to
override before the '=' sign otherwise it only overrides the tag for the image.
The previous version results in this entry (which does nothing to the manifest)

images:
- name: localhost:5000/cc-operator
  newTag: latest

and the new version does this:

images:
- name: quay.io/confidential-containers/operator
  newName: localhost:5000/cc-operator
  newTag: latest

The 'name' needs to be the quay entry and not 'controller' because there is
already a previous override in config/manager/kustomization.yaml.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>