crd: Drop RuntimeClassNames in favour of RuntimeClasses
RuntimeClassNames has been added as part of the commit
e0ab5aed4a865579ee6c0612bc6f43e23b944e66, but right now we need to
actually map a runtime with the snapshotter it'll be using, forcing us
to drop the RuntimeClassNames string from the CRD, and add a struct
which will hold the Name and the Snapshotter to be used.

This suggestion came from Pradipta Banerjee, and I appreciate a lot the
way of handling this, and I'm assuming here changes on the CRD are not
too problematic, as long as those are not done too often, while we're in
the v1beta1 stage.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>