pre-install-payload: Rename deploy script
The script had in its name "container-engine-for-cc-", which made sense
when we were only using it to deploy the container engine needed for cc.

However, we'll be expanding the functionality of the script to also make
sure the needed snapshotters are deployed at the users will.

With this in mind, let's simply rename the script to something more
generic like "reqs-deploy.sh".

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>