pre-install-payload: Rework Dockerfile
Let's start using a multi-step build Dockerfile in order to make things
a little bit more contained for each parts we'll build.

Together with this change, we're also moving to using alpine as the base
image, as it's way smaller than the Ubuntu one, and is also multi-arch.

I know, I know, it seems that having a multi-stage build here makes no
sense, as we'll end up using an alpine image in the end, but it'll pay
off by the moment we'll start building the artefacts, which for sure
will happen with the nydus snapshotter.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>