pre-install-payload: Rename COCO_${foo} to NODE_${foo}
We're renaming a few variables here to make sure we use COCO stuff only
for the projects we've forked, such as containerd.

This will help immensely when having to deal with the official and the
CoCo version forked of a specific component.

In the long run, of course, we want to get rid of all the forks we have,
but this is needed at least for this release.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>