pre-install-payload: Adapt script to specific containerd flavour
Let's change the reqs-deploy.sh script so it's clear that we're dealing
with a specific kind of containerd installation, such as the coco forked
one.  Later on we'll introduce an option to install an official
contained binary, just in case folks are running on an cluster that
doesn't have the minimum pre-requirements in place.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>