pre-install: config: Install nydus-snapshotter
The nydus-snapshotter is going to be used in order to replace our
dependency on the forked containerd.

For now, we're not actually removing the forked containerd dependency,
but rather giving the admin the choice to use nydus-snapshotter.

Using the forked containerd will be deprecated as part of our **next**
release.

Signed-off-by: ChengyuZhu6 <chengyu.zhu@intel.com>
Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>