pre-install: Temporarily use fidencio's nydus-snapshotter release
This is a release of nydus-snapshotter that has absolutely **ZERO** code
changed, but that adds a tarball for s390x.

This has been done in order to work around the issue we have seen
building the binaries for s390x as part of our CI, at least till
nydus-snapshotter releases a s390x tarball, which is being tracked as
part of https://github.com/containerd/nydus-snapshotter/issues/548
```

The error we're facing is:
22.67 -> unzip
   /go/pkg/mod/cache/download/github.com/freddierice/go-losetup/@v/v0.0.0-20220711213114-2a14873012db.zip:
zip: checksum error
22.69 go: downloading github.com/rs/xid v1.4.0
22.69 go: downloading github.com/containerd/stargz-snapshotter/estargz
   v0.14.3
22.69 go: downloading golang.org/x/net v0.10.0
22.71 go: downloading github.com/docker/distribution v2.8.2+incompatible
22.77 -> unzip
   /go/pkg/mod/cache/download/github.com/rs/xid/@v/v1.4.0.zip: zip:
checksum error
22.83 go: downloading github.com/moby/locker v1.0.1
24.73 pkg/filesystem/stargz_adaptor.go:19:2:
   github.com/KarpelesLab/reflink@v1.0.1: Get
"https://proxy.golang.org/github.com/%21karpeles%21lab/reflink/@v/v1.0.1.zip":
local error: tls: bad record MAC
```

This is not ideal, but this is "good enough" to unblock this release.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>