controllers: Workaround Uninstall / postUninstall race condition
This should be treated in a better way than adding an one minute sleep,
as just having the sleep here won't do us any good in the future.

What's basically happening, and forcing us to do this, is the fact that
the Uninstall and postUninstall daemonsets are being started at exactly
the same time, leading to a race condition when changing the containerd
configuration.

When looking at the kata-containers payload code, we see that the the
label is only set after containerd is successfully reconfigured, and
looking at this function we see we shouldn't reach this part before the
label is set.  However, that's not what we're facing ...

In order to unblock this PR to get merged, we're good enough.  But this
needs proper investigation and hopefully we'll see this patch being
reverted sooner than later.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>