pre-install: Create containerd config it it doesn't exist
We're relying here on containerd being on /usr/local/bin/, which we do
have access to, in order to create such config.

We're creating the config as the first thing on the script, as we don't
want to get into the situation where a configuration was added / removed
and we end up with that in the system due to us using a newer version of
containerd than the one present in the system.

We may face more issues in the future, such as:
* containerd is only in /usr/bin/
  * We can start exposing this as a host path mount, but let's not do it
    unless it becomes necessary
* when CRI-O is supported, this file should only be created in case
  containerd is used

However, for now, let's go with this approach and avoid users hitting a
possible error due to the lack of contained configuration.

Fixes: #271

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>