ccruntime: Set containerd / nydus-snapshotter env vars globally
In theory those could very well be set only for the pre-install /
post-uninstall, but that doesn't make much sense as the user experience
on setting those would be to having to set those twice.

With the user experience in mind, let's just move them to the global env
vars and let the user set / kustomize them only once.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>