reqs: Set the default value for the passed env vars
Otherwise we may hit issues in in cases where folks will use kustomize
to add / replace one the values, as kustomize just overrides the whole
set of env vars.

We need to keep those in sync with whatever is set in the configs, but
this shouldn't be as much of a burden as leaving it for the user to do
so.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>