pre-install: Remove nydus-snapshotter config
We've been leaving the nydus-snapshotter config behind because we were
trying to remove it from a path that was changed from the first
iteration of the nydus-snapshotter addition, but ended up not being
noticed during review.

While here, let's also make sure to entirely remove
`/opt/confidential-containers/share`, as nydus-snapshotter is the only
bit of code using that.

Fixes: #278

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>