reqs: handle the absence of disable_snapshot_annotations
The `disable_snapshot_annotations` should be set to `false` in the
containerd's configuration.toml in order to make the remote snapshotter
work. The reqs-deploy.sh currently handles only the case where the
property is already present on the configuration file, i.e., it will
switch from `true` to `false` only. However, on some nodes like in
AWS EKS and Azure AKS, `disable_snapshot_annotations` simply doesn't
exist so the current `sed` takes no effect (and the current containerd's
behavior is to disable annotations).

This changed the handler so that on the absence of
`disable_snapshot_annotations`, it is appended to the
`[plugins."io.containerd.grpc.v1.cri".containerd]` section. It will
leave a comment just below the added line so that the uninstall handler
know the line should be removed.

Signed-off-by: Magnus Kulke <magnuskulke@microsoft.com>
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>