config: add minKubeVersion to CSV
Supress the warning:

```
WARN[0000] Warning: Value : (cc-operator.v0.8.0) csv.Spec.minKubeVersion is not informed. It is recommended you provide this information. Otherwise, it would mean that your operator project can be distributed and installed in any cluster version available, which is not necessarily the case for all projects.
```

Kubernetes 1.24 is the version used on operator CI as well as the
mininum recommended version in the quickstart guide of CoCo.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>