tests: tagging the built operator image
The operator image is built then published on a local registry; on
config/manager/kustomization.yaml its name is changed to localhost:5000/cc-operator
so that it is pulled from that registry on install time. It isn´t setting the tag of
the image though, which means it uses whatever `newTag` value from the
kustomization file. Everything works fine when `newTag` is `latest`, but
when it is set to a release tag (e.g. v0.8.0) the installation will look for
an image like localhost:5000/cc-operator:v0.8.0, whereas on local
registry it was pushed as localhost:5000/cc-operator:latest. This
change ensures that `newTag` is always set to `latest`, matching the
image name and tag of the one pushed to the local registry.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>