gha: ccruntime_e2e: Avoid fail-fast
Every now and then we'll face errors that may be different depending on
the distro being used, due to some distro changes or even some flakiness
in the tests.

Till everything is 100% stable, which is not the case yet, we better
allow the tests to finish even if one test fails, giving us a better
idea of the problem being a generic one or specific to flakiness / one
distro.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>