gha: ccruntime_e2e: Add a 45 minutes timeout
Today I've faced one test hanging for more than one hour, stuck, without
giving us any useful information, leaving the "cancel" option (which is
up to the user to do) as the only option to finish the test and re-start
it.

As a possible way to avoid wasting resources for too long, let's simply
add a timeout of 45 minutes in the execution of the e2e tests.  This is
10 minutes more than what's been observed as the total time taken by
this step.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>