tests.e2e: Ensure we don't skip on kubectl failure
recent issues in CI indicate that kubectl might sometimes fail which
results in wait_for_process interrupting the loop. Let's improve the
command to ensure kubectl command passed and only then grep for the
(un)expected output.

Note the positive commands do not need to be treated as the output
should not contain the pod names on failure.

Fixes: #339

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>