crd: Expose agent{Https,No}Proxy
Depending on the environment where we deploy Confidential Containers,
setting up the proxy is required for the agent to be able to connect
with the external world.

With that in mind, mainly considering this is needed for the basic TDX
CI, let's ensure we expose to the users a way to set it up.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>