Release: attempt to fix docker manifest commands
This effectively reverts PR 366 and drops the guarding
quotes around extra_docker_manifest_flags that were added
to these commands in 351.
The CI is still throwing an error to the effect of:
"error parsing name for manifest list : invalid reference
format". This seems to stem from an empty string in
the command b/c of an empty extra_docker_manifest_flag
variable. Hence, we are reverting quotes that were added
around it in PR 351.

Signed-off-by: Chris Porter <porter@ibm.com>