release: fixup enclave-cc test
Enclave-cc now uses a "unified bundle" which means
that the paths to a few things have changed.

This was already accounted for in the cicd workflow,
but the e2e test was not modified.

Copy the change from the cicd workflow. We might
adopt a change without sed in the future.

Signed-off-by: Tobin Feldman-Fitzthum <tobin@ibm.com>