CI: Pin requests' version less than 2.32
The issue https://github.com/docker/docker-py/issues/3113 seems to be resolved.
The pinned version of requests (2.28.1) is no longer compatible with the current
docker pip, resulting in errors like:

"Error connecting: Error while fetching server API version: Not supported URL scheme http+docker"

Initially, we attempted to allow pip to install requests according to its dependency tree.
However, the issue persists on Ubuntu 22.04, and it appears to be related to
https://github.com/docker/docker-py/issues/3256.
The best solution so far is to pin the version of the package to less than 2.32.

Fixes: #381

Signed-off-by: Hyounggyu Choi <Hyounggyu.Choi@ibm.com>