controllers: Introduce cc-startuninstall label to signal uninstallation
currently the operator uses katacontainers.io/kata-runtime=cleanup label
to start the daemon uninstall as well as to check the kata uninstall
process is done. This commit introduces a new cc-startuninstall label
which is set by this operator and used to run the uninstall DS while it
keeps using the katacontainers.io/kata-runtime=cleanup label to detect
kata cleanup is finished.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>