tests: Use ccruntime status to check the operator is installed
currently the tests/e2e/operator.sh checks for various pods presense as
well as for runtimeclasses to be present to declare the operator is
installed.

As this function is also used to ensure the operator is
somehow-functional let's keep the runtimeclass check and decrease the
wait-time (theoretically it should be already present but due to caching
it might take a few seconds to be correctly reported).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>