tests: Send -9 on cleanup interruption
we are using "timeout" to limit cleanup time in CI, but the default -15
won't interrupt on some occasions. Let's use -9 instead to not to block
in case the cleanup fails.

Note this is only about the cleanup part of the e2e run-local.sh, every
hang inside the testing itself should still result in a failure.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>