runners: Free space on the GitHub runners
Those runner usually have ~14GB free for utilization, which is rather
low, mainly when we take into consideration the size of the payloads we
have (due the recent switch to zst, on the Kata Containers side).

With this said, let's take the approach already taken by peer-pods (and
thanks Steve for the tip :-)), and free as much space as possible before
running our tests.

Signed-off-by: Fabiano Fidêncio <ffidenciodeo@nvidia.com>